Added new component
-FormImagePicker component added
- Code cleanup
