+++ app/components/forms/AppFormPicker.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { useFormikContext } from "formik";
 import AppPicker from "../AppPicker";
-import { ErrorMessage } from "./ErrorMessage";
 
 function AppFormPicker({
   items,
@@ -14,15 +13,17 @@ function AppFormPicker({
   const { errors, setFieldValue, touched, values } = useFormikContext();
 
   return (
-    <AppPicker
-      items={items}
-      onSelectItem={(item) => setFieldValue(name, item)}
-      placeholder={placeholder}
-      selectedItem={values[name]}
-      width={width}
-      PickerItemComponent={PickerItemComponent}
-      numberOfColumns={numberOfColumns}
-    />
+    <>
+      <AppPicker
+        items={items}
+        onSelectItem={(item) => setFieldValue(name, item)}
+        placeholder={placeholder}
+        selectedItem={values[name]}
+        width={width}
+        PickerItemComponent={PickerItemComponent}
+        numberOfColumns={numberOfColumns}
+      />
+    </>
   );
 }
